refactor: qora, react-toastify cdn
